requirements.txt
pydantic==1.9.1
websockets==10.3
pynput==1.7.6
python-decouple==3.6
